Add Kafka Producer for Order Notification
This commit introduces the following changes to the codebase:

- Add `spring-kafka` dependency to `pom.xml`
- Create `KafkaConfig` class to define a `NewTopic` bean for `notificationTopic`
- Create `OrderNotification` class to encapsulate order number
- Update `OrderService` to send `OrderNotification` to `notificationTopic` using `KafkaTemplate` after order creation
- Add Kafka producer properties in `application.properties`

Purpose:
The purpose of this commit is to integrate Kafka into the Order Service application for publishing order notification events. After an order is successfully created, an `OrderNotification` object containing the order number is sent to the `notificationTopic` Kafka topic using the `KafkaTemplate`. This allows other services or consumers to subscribe to this topic and receive order notifications for further processing, such as sending notifications to customers or triggering additional workflows.